Adding types and getting cost working
