+++ internal/providers/pulumi/aws/util.go
@@ -1,0 +1,13 @@
+package aws
+
+func intPtr(i int64) *int64 {
+	return &i
+}
+
+func strPtr(s string) *string {
+	return &s
+}
+
+func floatPtr(f float64) *float64 {
+	return &f
+}
