+++ internal/providers/pulumi/parser.go
@@ -1,74 +1,15 @@
 package pulumi
 
 import (
+	"encoding/json"
 	"fmt"
 	"strings"
-	"time"
 
 	"github.com/infracost/infracost/internal/config"
+	"github.com/infracost/infracost/internal/providers/pulumi/types"
 	"github.com/infracost/infracost/internal/schema"
-	"github.com/pulumi/pulumi/pkg/v3/engine"
-	"github.com/pulumi/pulumi/pkg/v3/resource/deploy"
-	"github.com/pulumi/pulumi/sdk/v3/go/common/apitype"
-	"github.com/pulumi/pulumi/sdk/v3/go/common/diag"
-	"github.com/pulumi/pulumi/sdk/v3/go/common/resource"
 	"github.com/tidwall/gjson"
 )
-
-// previewDigest is a JSON-serializable overview of a preview operation.
-type PreviewDigest struct {
-	// Config contains a map of configuration keys/values used during the preview. Any secrets will be blinded.
-	Config map[string]string `json:"config,omitempty"`
-
-	// Steps contains a detailed list of all resource step operations.
-	Steps []*previewStep `json:"steps,omitempty"`
-	// Diagnostics contains a record of all warnings/errors that took place during the preview. Note that
-	// ephemeral and debug messages are omitted from this list, as they are meant for display purposes only.
-	Diagnostics []previewDiagnostic `json:"diagnostics,omitempty"`
-
-	// Duration records the amount of time it took to perform the preview.
-	Duration time.Duration `json:"duration,omitempty"`
-	// ChangeSummary contains a map of count per operation (create, update, etc).
-	ChangeSummary engine.ResourceChanges `json:"changeSummary,omitempty"`
-	// MaybeCorrupt indicates whether one or more resources may be corrupt.
-	MaybeCorrupt bool `json:"maybeCorrupt,omitempty"`
-}
-
-// propertyDiff contains information about the difference in a single property value.
-type propertyDiff struct {
-	// Kind is the kind of difference.
-	Kind string `json:"kind"`
-	// InputDiff is true if this is a difference between old and new inputs instead of old state and new inputs.
-	InputDiff bool `json:"inputDiff"`
-}
-
-// previewStep is a detailed overview of a step the engine intends to take.
-type previewStep struct {
-	// Op is the kind of operation being performed.
-	Op deploy.StepOp `json:"op"`
-	// URN is the resource being affected by this operation.
-	URN resource.URN `json:"urn"`
-	// Provider is the provider that will perform this step.
-	Provider string `json:"provider,omitempty"`
-	// OldState is the old state for this resource, if appropriate given the operation type.
-	OldState *apitype.ResourceV3 `json:"oldState,omitempty"`
-	// NewState is the new state for this resource, if appropriate given the operation type.
-	NewState *apitype.ResourceV3 `json:"newState,omitempty"`
-	// DiffReasons is a list of keys that are causing a diff (for updating steps only).
-	DiffReasons []resource.PropertyKey `json:"diffReasons,omitempty"`
-	// ReplaceReasons is a list of keys that are causing replacement (for replacement steps only).
-	ReplaceReasons []resource.PropertyKey `json:"replaceReasons,omitempty"`
-	// DetailedDiff is a structured diff that indicates precise per-property differences.
-	DetailedDiff map[string]propertyDiff `json:"detailedDiff"`
-}
-
-// previewDiagnostic is a warning or error emitted during the execution of the preview.
-type previewDiagnostic struct {
-	URN      resource.URN  `json:"urn,omitempty"`
-	Prefix   string        `json:"prefix,omitempty"`
-	Message  string        `json:"message,omitempty"`
-	Severity diag.Severity `json:"severity,omitempty"`
-}
 
 type Parser struct {
 	ctx *config.ProjectContext
@@ -114,7 +55,7 @@ func (p *Parser) createResource(d *schema.ResourceData, u *schema.UsageData) *sc
 	}
 }
 
-func (p *Parser) parsePreviewDigest(t *PreviewDigest, usage map[string]*schema.UsageData, rawValues gjson.Result) ([]*schema.Resource, []*schema.Resource, error) {
+func (p *Parser) parsePreviewDigest(t types.PreviewDigest, usage map[string]*schema.UsageData, rawValues gjson.Result) ([]*schema.Resource, []*schema.Resource, error) {
 	baseResources := p.loadUsageFileResources(usage)
 
 	var resources []*schema.Resource
@@ -129,6 +70,10 @@ func (p *Parser) parsePreviewDigest(t *PreviewDigest, usage map[string]*schema.U
 		var name = step.NewState.URN.Name().String()
 		var resourceType = step.NewState.Type.String()
 		var providerName = strings.Split(step.NewState.Type.String(), ":")[0]
+		var localInputs = step.NewState.Inputs
+		localInputs["config"] = t.Config
+		var inputs, _ = json.Marshal(localInputs)
+
 		tags := map[string]string{} // TODO: Where do I get tags?
 		var usageData *schema.UsageData
 
@@ -136,8 +81,7 @@ func (p *Parser) parsePreviewDigest(t *PreviewDigest, usage map[string]*schema.U
 			usageData = ud
 		}
 
-		resourceData := schema.NewResourceData(resourceType, providerName, name, tags, rawValues)
-
+		resourceData := schema.NewResourceData(resourceType, providerName, name, tags, gjson.Parse(string(inputs)))
 		if r := p.createResource(resourceData, usageData); r != nil {
 			if step.Op == "same" {
 				pastResources = append(resources, r)
